requirements.txt
Django==1.11
django-bootstrap3==8.2.3
django-braces
django-autoslug==1.9.3
django-debug-toolbar==1.8
djangorestframework==3.6.3